refactor: clean up and rename randomQuoteReducer.js to reducer.js
